Moved the put survey answers in the survey functions file
